Use qd-library for double-double arithmetic.
Rather than using `long double`, which is not portable beyond x86 on
Linux and Mac, use the qd-library for double-double arithmetic.  This
has the advantage of greater precision than `long double` as well.  A
bundled version is used by default, or the system copy may be used by
passing `--use-system-qd`.