Fix computation of intersection of disjoint spherical polygons
The intersection and union of spherical polygons can result in
disjoint polygonss. This package represents them as lists of simple
polygons and performs further operations on them by combining the
results of the simple polygons which comprise them. The code which
computed the intersection of disjoint polygons was incorrect. It
combined the two lists of simple polygons and performed the
intersection between all of them. However, this is obviously
incorrect, because the the intersection of disjoint polygons is always
the empty polygon. This update modifies the intersection code so that
it computes the intersection between each possible pair of simple
polygons and takes the union of all the results.

The code for multi_intersection was modified to call intersection. It
simply calls intersection on each spherical polygon and the previous
result, which is initialized to the first polygon. The optional
parameter method was removed, as it has no meaning in the new code.