Modify the area computation to use inside polygon point
The previous area computation assumed that the polygon was smaller
than a hemisphere (2 pi) and converted the computed area to a negative
number when larger. The new code looks at the polygon vertices and if
they are ordered clockwise with respect to the inside point, the area
is poitive and if they are ordered counter-clockwise, the area is
negative. This change allows all sizes of polygon to be handled, up to
4 pi. To support this, a new method, is_clockwise has been added to
_SingleSphericalPolygon.

I also added auto_close and auto_orient parameters when creating new
polygon objects. Both are false by default. If auto_close is set to
True, the first point in the polygon is added as the last, closing the
polygon. If auto_orient is True and the polygon vertices are ordered
counter-clockwise, the list of vertices is reversed to orient them
clockwise.

Testing revealed that the computation of the internal point was
incorrect for the special cose of a triangle, so I rewote that code. I
also modified the general case to choose the vertex with the sharpest angle.