Modify vector_to_lonlat to handle numpy arrays without throwing error
This is a modification to a previous fix, which shifted the range of
longitude from [-pi, pi] to [0, 2*pi]. The former is the range of the
atan2 function used to compute longitude (lon). The previous fix used
an if test with the less than operator, which does not work if the
inputs are numpy array. The new code uses the numpy remainder function,
which works with scalar inputs as well as numpy arrays.