Add a new step to union computation to remove degenerate edges
The graph used to compute the union of several polygons can contain
edges where both endpoints are the same. I call these degenerate
edges. These degenerate edges can lead to false positives in the step
which removes 3-connected edges. When edges are removed that shouldn't
be, the result is a disconnected outline for the union. I add a new
step to look for degenerate edges and remove them before calling the
step that removes 3-connected edges,