Fix inconsistent results from _find_all_intersections
The union and intersection methods on polygons return results that
vary from run to run. The source of the problem is in the code called
from _find_all_intersections. That code traverses the set of edges and
splits them where they intersect. The old edge is deleted and replaced
by the two edges formed by the intersection. The new edges are also
searched for intersections.

One of the features of python sets is that the order they are
traversed in differs from run to run, as the underlying data structure
is a hash. This leads to inconsistent results from run to run, as
edges are checked in a different order. While, ideally I would like to
track down the cause of the inconsistency, I haven't been able to do
so.  So I have made two changes which are of the nature of a
patch. First, I now fix the order in which the edges are checked by
sorting them. This required adding two short functions, whcih define an
(arbitrary) sort order. Second, I place the new edges produced by the
intersection at the end of the list to be checked instead of at the
beginning. This reduces the "butterfy effect" that arises if an
intersection is found on one run instead of another. In addition, I
did a small amount of cleanup by removing list coercions to the set of
edges where it was not necessary. This last change should have no
effect on the results.