Fix problem with adding new edges in _find_arc_to_arc_intersections
Part of the previous fix was to add new edges derived from the
intersection at the end of the list instead of at the beginning. There
were several places where this change needed to be made and I missed
one, leading to errors in the test results. So I fixed the remaining
line.