Make numpy.linspace step count an integer
The numpy function linspace computes an evenly spaced array of
values. In the current version of numpy it checks if the number of
steps is an integer.
Method from_code computes a spherical polygon representing the
intersection of a cone with the unit sphere. The number of steps used
in computing the pyramid approximating the cone had a default value
that was a float. That value was used when calling linspace, resulting
in an error. So the value has been changed from a float to an integer.
Get_point_set is a utility function used in testing. It covers an
ra/dec box in a grid of points. The number of points is less at higher
latitudes, multiplied by the cosine of the declination. This number is
passed to linspace, also triggering an error. This number is now
converted to an integer before calling linspace.