Update the setup files and documentation for the package
The travis file was updated to better match the astropy template. I
updated the requirements in the README file file and removed the cut
line section and replaced it on a section on disjoint polygons to
reflect how the code now works. I fixed the github project name in the
setup config file and bumped the version number in the stup file.