Change the order of the steps when creating a new polygon
A new spherical polygon should have its vertices in clockwise order
and an interior point defined. The __init__ function has code to
optionally re-order the vertices and to compute an interior point if
one is not defined. The computation of the interior point should be
first because the determination of whether the vertices are clockwise
depends on it. However, the computation of the interior point also
depends on whether the vertices are clockwise, creating a dilemma. Now
the __init__ function computes the interior point, then the ordering,
and then recomputes the interior point if the vertices are
re-ordered. This works because the computation of the interior point
falls back on a simpler method if the points are counter-clockwise.

I also made the method is_clockwise a method of SphericalPolygon,
mostly for my debugging, but also because, why not? It can be useful.