Re-order code in _SingleSphericalPolygon.__init__
When creating a spherical polygon one has the option of asking it to
sort the vertices so they are clockwise. This clockwise order depends
upon the location of the interior point. The code also has the option
of automatically computing the interior point. That depends on the
points being in clockwise order. This creates a dilemma in the case
you ask the code to do both. The solution in the rewritten code is to
first to compute the interior point, then re-order the vertices, and
then recompute the interor point. This ensures both calculations are
consistent with each other.