Re-order code in _SingleSphericalPolygon.__init__ (#114)
* This fix addresses issue #113

The travis file was updated to better match the astropy template. I
updated the requirements in the README file file and removed the cut
line section and replaced it on a section on disjoint polygons to
reflect how the code now works. I fixed the github project name in the
setup config file and bumped the version number in the stup file.

* Restore preamble of .travis.yml file

In an attempt to fix the travis error.

* Change the order of the steps when creating a new polygon

A new spherical polygon should have its vertices in clockwise order
and an interior point defined. The __init__ function has code to
optionally re-order the vertices and to compute an interior point if
one is not defined. The computation of the interior point should be
first because the determination of whether the vertices are clockwise
depends on it. However, the computation of the interior point also
depends on whether the vertices are clockwise, creating a dilemma. Now
the __init__ function computes the interior point, then the ordering,
and then recomputes the interior point if the vertices are
re-ordered. This works because the computation of the interior point
falls back on a simpler method if the points are counter-clockwise.

I also made the method is_clockwise a method of SphericalPolygon,
mostly for my debugging, but also because, why not? It can be useful.

* Re-order code in _SingleSphericalPolygon.__init__

When creating a spherical polygon one has the option of asking it to
sort the vertices so they are clockwise. This clockwise order depends
upon the location of the interior point. The code also has the option
of automatically computing the interior point. That depends on the
points being in clockwise order. This creates a dilemma in the case
you ask the code to do both. The solution in the rewritten code is to
first to compute the interior point, then re-order the vertices, and
then recompute the interor point. This ensures both calculations are
consistent with each other.