Remove _remove_cut_lines
Intersection and union of polygons can lead to disjoint
polygons. Formerly this were represented as a single polygon joined by
cut lines, polygon sidea containing zero area. The code was changed a
while back to represent disjoint polygons as lists of polygons. So the
code to handle cut lines during unions and intersections is no longer
needed and this update removes it. The change was prompted by an error
in computing the intersection of two polygons with a shared side. The
shared side was mistaken for a cut line, leading to a mistake in the
result.