Move cut line removal routine in union and intersection
A user submitted a problem report when computing the overlap between
two polygons where one polygon was contained in the other and both
shared a common edge. The code said that the overlap was zero, when it
should have been one. The problem was traced to the cut line removal
code. There is no clean way to distinguish between a cut line and a
shared edge, as in both cases there are two connected nodes with three
edges. The spherical geometry code no longer creates polygons with cut
lines. They were a method previously used to represent disjoint
polygons. But since they are sill, encountered, at least in the test
code, the code to check for and remove them cen't be eliminated. But
the step that removes them can be moved downstream of the calculation
of interior or exterior points. This resolves the problem of mistaking
a shared edge for a cut line and the reported problem.