Change polygons defined with cut lines into disjoint polygons
Older versions of spherical_geometry defined disjoint polygons as a
single polygon joined by cut lines. When taking the intersection or
union between polygons, the first step was to remove thes cut lines.
However, when two polygons have a shared edge, the edge looks like a
cut line and is removed, causing an error with the intereection and
union methods. This fix moves the removal of the cut lines into the
initialization of a SphericalPolygon. The single polygon with cut
lines is thus transformed into a list of polygons without cut
lines. And the initial step of removing cut lines in the union and
intersection methods can be removed, which fixes the problem of the
spurious removal of the shared edge of two polygons.