Remove dependency on astropy six library
As part of its move to drop support for Python 2, astropy will no
longer include the six library. There is one dependency on this
library in graph.py, in the _remove_cut_lines method. That method has
been rewritten to resemble the other _remove methods and no longer
uses the six library.