Rewrite is_clockwise test and other changes (#127)
This update rewrites the is_clockwise test in polygon.py. A number of
other changes to the package follow from that change. The motivation
was a test case where a user created a plygon whose vertices fall on
lines of constant latitude and longitude. Lines of constant longitude
are great circles, but lines of constant latitude are not, except for
the equator. Since spherical polygon sides are great circles, this
creates a scalloped polygon woth a mix of convex and concave
angles. The previous is_clockwise test incorrectly handled this kind
of polygon. It also caused errors in the calculation of the inside
point in the case none was specified when creating a spherical polygon.

The new code in is_clockwise calculates whether the two vectors
forming the sides of a vertex and the vertex vector form a left handed
or right handed system. This depends on the order the two sides are
traversed. One way will give a right handed system and the opposite a
left handed system. The handedness is given by the sign of the scalar
triple product of the three vectors. Summing the scalar triple product
at all the vertices gives the overall handedness. If right handed, the
polygon is ordered counter-clockwise. If left-handed, the polygon is
ordered clockwise.

So this update adds a new function to compute the scalar triple
product to great_circle_arc. Future plans are to use this code to test
for co-linearity at several places, as well as to write a C version
that can be used by numpy.

The formula for a polygon area depends upon the orientation of the
polygon (clockwise or counter-clockwise) but not on the inside
point. The new clockwise test also does not depend on the inside
point. Since the orientation is critical for getting correct results
from this package, I have modified the code to always re-orient a
polygon so that it is clockwise if the inside point is in the smaller
portion of the sphere or counter-clockwise if it is in the larger
portion of the sphere. This is done when creating a new spherical
polygon. If no inside point is given, the code calculates an inside
point in the smaller portion. This re-orientation was only done
previously when auto_orient was set to True. Since this parameter is
redundant, I have removed it.

One consequence of this change is that now the area of a polygon is
always a positive number between zero and four pi. There are no more
negative areas. This is because the re-orientation of polygons means
that the orientation never disagrees with the location of the inside
point.