Revised code for choosing between edges when tracing polygons (#130)
When taking the union or intersection of two polygons, the result is a
set of edges. The code then traces a closed path through the edges one
or more times to construct yje polygons. There are cases where more
than one edge is connected to a a previous edge. In this case we want
to choose an edge that does no create a self-intersecting
polygon. There was an error in the code which caused the first edge to
always be selected. The new code choose an edge which makes an angle
with the previous edge that is midmost of the choices. This avoids
choosing an edge that makes a straight line with the previous edge.

I also removed use of same_arc for the present, which seems to have a
rounding problem.