Better handling for self-intersecting polygons
The last update introduced support for creating polygons from
self-intersecting polygons. It added code to check for intersections
and generate the list of simple, disjoint polygons. However, not all
the methods that create new polygons called the checking code. This
update modifies the polygon creating methods to call the new code,
which is in the __init__ method of SpherivalPolygon.

Testing found that the intersection checking code could create
unclosed polygons, where the first and final points are not the
same. So the code that closes a polygon is now always called and the
option which used to control it, auto_close has been removed. Since
this is a change to the interface, the version number has been changed
to 1.2.

The test suite reported that the numpy function expand_dims was no
longer supporting the way the spherical geometry package was using it,
adding a second dimension of length one to a one dimensional vector. I
added a new function, two_d to vector.py which calls numpy reshape and
does the same work expand_dims used to do.

I deleted the functions same_point and same_arc, as they were not
being used by the other code and sometimes generated wrong results. It
looks like the quad precision math routines has removed the need to
check for points that are nearly the same but not identical.