Revise Travis file and rewrite method to compute outside point
The Travis file was revised to be in better conformance with the
current astropy library template. The function that computes a new
outside point failed one test, so I rewrote it using a different
algorithm. However, I'm not completely happy with it and may revisit
the problem at a later date.