Rewrite invert_polygon to avoid contains_point check
The contains_point check appears to be a little flakey. I will
investigate that later, but for now I have rewritten the
invert_polygon code and _find_new_outside so that they don't perform
this check.

The C versions of the functions in great_circle_arc were accidentally
left turned off, for devugging. This patch turns them on again.