Revise Travis file and rewrite method to compute outside point (#136)
The Travis file was revised to be in better conformance with the
current astropy library template. The function that computes a new
outside point failed one test, so I rewrote it using a different
algorithm. The contains_point check appears to be a little flakey. I will
investigate that later, but for now I have written the
invert_polygon code and _find_new_outside so that they don't perform
this check.