Remove debugging code from graph.py (#149)
Spherical geometry creates a graph structure to perform intersections
ans unions. This code includes assertions to check that the graph
structure is well formed through each operation of the intersection
and union. Previously the code plotted the graph structure when the
assertions failed. I've removed the code because it required the
installation of other libraries to perform the plot. We can always add
whatever extra code we think is necessary locally, but other sites
shouldn't be forced to install libraries that we use for our
debugging. This fix addresses issue #2764 in the jwst package, which
uses spherical_geometry.